Upgrade Vector to RRB Vector.
RRB Vectors are an extension of the old Vector that allows the
implementation of concatenation on top of the other operation
with minimal loss of performace of other operations. It also
also allows implementations of insertion (of element or vectors)
at some index and remove element at index or between range.
This also improve the performace of parallel vector combiners.

Extensive documentation and benchmark results can be found in the
'Turning Relaxed Radix Balanced Vector from Theory into Practice for Scala Collections'
master thesis (done in the LAMP) and in the 'RRB Vector - A Practical
General Purpose Immutable Sequence' ICFP paper.